my_carwash/requirements.txt
appdirs==1.4.3
Django==1.10.4
django-filter==1.0.2
django-jet==1.0.5
djangorestframework==3.6.2
google-api-python-client==1.5.5
httplib2==0.10.3
Markdown==2.6.8
mysqlclient==1.3.10
oauth2client==4.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
pytz==2017.2
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0